corrosion-backend/requirements.txt
torch==1.12.0
torchvision==0.13.0
flask==2.2.2
flask-cors==3.1.1
Pillow==9.0.1
opencv-python==4.5.5.64
gdown==4.4.0
